igrab: first attempt at building

+++ source/igrab/xms.h
@@ -1,8 +1,8 @@
 int InitXMS(void);
-int XMSAllocate(long size);
+int XMSAllocate(int32_t size);
 unsigned XMSTotalFree(void);	// returns KB free
 void XMSFreeMem(int handle);
-void XMSmove(int srchandle,long srcoff,int desthandle,long destoff,long size);
+void XMSmove(int srchandle,int32_t srcoff,int desthandle,int32_t destoff,int32_t size);
 void XMSHandleInfo(int handle);
 
 extern unsigned XMSavail;
